requirements.txt
beautifulsoup4
fastapi
matplotlib
numpy
openai
praw
pydantic
python-dotenv
python_dateutil
pytz
Requests
scikit_learn
sentence_transformers
SQLAlchemy
SQLAlchemy_Utils
transformers
yfinance
psycopg2
redis